+++ Jolt/ObjectStream/ObjectStreamTextIn.h
@@ -27,7 +27,7 @@ public:
 	virtual bool				ReadPrimitiveData(int &outPrimitive) override;
 	virtual bool				ReadPrimitiveData(uint32 &outPrimitive) override;
 	virtual bool				ReadPrimitiveData(uint64 &outPrimitive) override;
-	virtual bool				ReadPrimitiveData(float &outPrimitive) override;
+	virtual bool				ReadPrimitiveData(decimal &outPrimitive) override;
 	virtual bool				ReadPrimitiveData(double &outPrimitive) override;
 	virtual bool				ReadPrimitiveData(bool &outPrimitive) override;
 	virtual bool				ReadPrimitiveData(String &outPrimitive) override;
